pyproject.toml
[project]
name = "project-3"
version = "0.1.0"
description = ""
authors = [
    {name = "n-emelianov",email = "limbolume2023@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12, <3.13"
dependencies = [
    "uvicorn (>=0.41.0,<0.42.0)",
    "fastapi (>=0.135.1,<0.136.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "python-dotenv>=1.0.0,<2.0.0",
    "psycopg2-binary>=2.9.0,<3.0.0",
    "redis>=5.0.0,<6.0.0",
    "asyncpg (>=0.31.0,<0.32.0)",
    "setuptools>=69.0.0",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "aiosqlite (>=0.22.1,<0.23.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "greenlet (>=3.3.2,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "coverage (>=7.13.5,<8.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "locust (>=2.43.3,<3.0.0)"
]
